feat: Convert TimePilot game codebase from AMD to ESM TypeScript
- Added main game logic in `src/game/index.ts`
- Implemented menu system in `src/game/menus.ts` and related files
- Created player functionality in `src/game/player.ts`
- Developed prop and prop factory in `src/game/prop.ts` and `src/game/prop-factory.ts`
- Defined types in `src/game/types.ts` for better type safety
- Updated user options management in `src/game/user-options.ts`
- Enhanced HTML test file for better structure and readability
- Updated TypeScript build configuration to reflect new file structure
