+++ src/game/controller/keyboard2.ts
@@ -1,8 +1,8 @@
-/* Converted from TimePilot.Controller.Keyboard1.js (AMD) to ESM TypeScript. */
-import helpers from "./engine/helpers";
-import type { Controller, ControllerInterfaceInstance } from "./TimePilot.types";
+/* Converted from TimePilot.Controller.Keyboard2.js (AMD) to ESM TypeScript. */
+import helpers from "../engine/helpers";
+import type { Controller, ControllerInterfaceInstance } from "../types";
 
-var Keyboard1 = function (controllerInterface: ControllerInterfaceInstance) {
+var Keyboard2 = function (controllerInterface: ControllerInterfaceInstance) {
   this._controllerInterface = controllerInterface;
 
   this.connect();
@@ -11,7 +11,7 @@ var Keyboard1 = function (controllerInterface: ControllerInterfaceInstance) {
   prototype: Record<string, unknown>;
 };
 
-Keyboard1.prototype = {
+Keyboard2.prototype = {
   connect: function () {
     var that = this;
     helpers.bind(
@@ -21,22 +21,12 @@ Keyboard1.prototype = {
           case 37: // Left-Key
           case 65: // "A"
             event.preventDefault();
-            that._controllerInterface.rotateToHeading(270);
-            break;
-          case 38: // Up-Key
-          case 87: // "W"
-            event.preventDefault();
-            that._controllerInterface.rotateToHeading(0);
+            that._controllerInterface.rotateAntiClockwise();
             break;
           case 39: // Right-Key
           case 68: // "D"
             event.preventDefault();
-            that._controllerInterface.rotateToHeading(90);
-            break;
-          case 40: // Down-Key
-          case 83: // "S"
-            event.preventDefault();
-            that._controllerInterface.rotateToHeading(180);
+            that._controllerInterface.rotateClockwise();
             break;
           case 32: // Space-Bar
             event.preventDefault();
@@ -48,16 +38,12 @@ Keyboard1.prototype = {
             break;
           case 27: // Escape-Key
             event.preventDefault();
-            that._controllerInterface.toggleMenu();
+            that._controllerInterface.openMenu();
             that._controllerInterface.togglePause();
             break;
           case 80: // "P"-Key
             event.preventDefault();
             that._controllerInterface.togglePause();
-            break;
-          case 82: // "R"-Key
-            event.preventDefault();
-            that._controllerInterface.restart();
             break;
         }
       },
@@ -71,17 +57,12 @@ Keyboard1.prototype = {
           case 27: // Escape-Key
           case 70: // "F"-Key
           case 80: // "P"-Key
-          case 82: // "R"-Key
             event.preventDefault();
             break;
           case 37: // Left-Key
-          case 38: // Up-Key
           case 39: // Right-Key
-          case 40: // Down-Key
-          case 65: // "A"-Key
-          case 87: // "W"-Key
-          case 68: // "D"-Key
-          case 83: // "S"-Key
+          case 65: // "A"
+          case 68: // "D"
             event.preventDefault();
             that._controllerInterface.stop();
             break;
@@ -101,4 +82,4 @@ Keyboard1.prototype = {
   },
 };
 
-export default Keyboard1;
+export default Keyboard2;
