+++ src/game/data-store.ts
@@ -1,5 +1,5 @@
 /* Converted from TimePilot.dataStore.js (AMD) to ESM TypeScript. */
-import type { GameDataStore } from "./TimePilot.types";
+import type { GameDataStore } from "./types";
 
 var dataStore = {} as GameDataStore;
