+++ src/game/menus/main.ts
@@ -1,5 +1,5 @@
 /* Converted from TimePilot.Menu.main.js (AMD) to ESM TypeScript. */
-import type { MenuDefinition } from "./TimePilot.types";
+import type { MenuDefinition } from "../types";
 
 var main: MenuDefinition = {
   name: "Welcome",
