+++ test.html
@@ -1,41 +1,46 @@
-<!DOCTYPE html>
+<!doctype html>
 <html>
-<head>
-    <meta charset='utf-8' />
+  <head>
+    <meta charset="utf-8" />
 
     <title>TimePilot</title>
 
     <script src="./lib/require.min.js"></script>
     <script>
-        require({
-            baseUrl: "./"
-        }, [
-            "TimePilot"
-        ], function (TimePilot) {
-            new TimePilot(document.getElementById("test"), {
-                debug: true
-            });
+      require({
+        baseUrl: "./",
+      }, ["TimePilot"], function (TimePilot) {
+        new TimePilot(document.getElementById("test"), {
+          debug: true,
         });
+      });
     </script>
     <style>
-        html,
-        body {
-            margin: 0;
-            padding: 0;
-        }
-        #test {
-            width: 720px;
-            height: 480px;
-            background-color: #F88
-        }
-        #test canvas {
-            background-color: #FD8
-        }
+      html,
+      body {
+        margin: 0;
+        padding: 0;
+      }
+      #test {
+        width: 720px;
+        height: 480px;
+        background-color: #f88;
+      }
+      #test canvas {
+        background-color: #fd8;
+      }
     </style>
-</head>
+  </head>
 
-<body>
-    <div id='test'></div>
-    <script>document.write('<script src="http://' + (location.host || 'localhost').split(':')[0] + ':35729/livereload.js?snipver=1"></' + 'script>')</script>
-</body>
+  <body>
+    <div id="test"></div>
+    <script>
+      document.write(
+        '<script src="http://' +
+          (location.host || "localhost").split(":")[0] +
+          ':35729/livereload.js?snipver=1"></' +
+          "script>"
+      );
+    </script>
+  </body>
 </html>
